Add conversion logic to rates adapter

+++ app/src/main/java/com/bobowiec/revolut_app/data/model/CurrentRates.kt
@@ -1,3 +1,3 @@
 package com.bobowiec.revolut_app.data.model
 
-data class CurrentRates(val base: String, val date: String, val rates: Map<String, Float>)
+data class CurrentRates(val base: String, val date: String, val rates: Map<String, Double>)
